Only reset highlightedIndex on input value change
The previous approach was causing highlightedIndex to be cleared for every
keystroke, even those that didn't cause the input value to change (ctrl, cmd,
arrow keys, etc). This was especially frustrating when using the left/right
arrow keys to move the cursor, as it would cause the highlight to disappear
until the user used the up/down arrow keys or changed the input value.