Prevent onFocus and onBlur when selecting an item #229
inputProps.onFocus and inputProps.onBlur were incorrectly being invoked when
using the mouse to select an item from the menu. This is incorrect because
although a focus and blur actually occur, the user does not perceive it. The
Autocomplete internals ensure focus is restored after the user has clicked
outside the <input> to select the item.